+AuthService, в UserQueries только обращения к базе, res -> user
