+++ lib/twitter_web/controllers/users_controller.ex
@@ -14,7 +14,7 @@ defmodule TwitterWeb.UsersController do
   def show(conn, _params) do
     user = Guardian.Plug.current_resource(conn)
     IO.inspect(user)
-    conn |> render("user.json", res: user)
+    conn |> render("user.json", user: user)
  end
 
   def sign_in(conn, %{"email" => email, "password" => password}) do
