+++ lib/twitter_web/views/users_view.ex
@@ -5,7 +5,7 @@ defmodule TwitterWeb.UsersView do
     %{jwt: jwt}
   end
 
-  def render("user.json", %{res: res}) do
-    %{email: res.email, username: res.username}
+  def render("user.json", %{user: user}) do
+    %{email: user.email, username: user.username}
   end
 end
